dont overwrite files in generators
problem: 
- current generator is overwriting files without questions

solution: 
- for now skip the generation, if a file was already found

+++ lib/rdm/gen/init.rb
@@ -55,9 +55,14 @@ module Rdm
 
       private
 
+      def warning(msg)
+        puts Rdm::Support::Colorize.brown(msg)
+      end
+
       def ensure_file(path_array, content = '')
         filename = File.join(*path_array)
         FileUtils.mkdir_p(File.dirname(filename))
+        return warning("File #{filename} already exists, skipping...") if File.exist?(filename)
         File.write(filename, content)
       end
 
@@ -65,6 +70,7 @@ module Rdm
         from          = filepath
         target_name ||= filepath
         to            = File.join(current_dir, target_name)
+        return warning("File #{to} already exists, skipping...") if File.exist?(to)
         FileUtils.mkdir_p(File.dirname(to))
         # copy_entry(src, dest, preserve = false, dereference_root = false, remove_destination = false)
         FileUtils.copy_entry(from, to, true, false, true)
