+++ lib/rdm/gen/package.rb
@@ -79,16 +79,20 @@ module Rdm
 
       def move_templates
         Dir.chdir(templates_path) do
-          copy_template('.gitignore')
           copy_template('bin/console_irb', 'bin/console')
         end
       end
 
       private
 
+      def warning(msg)
+        puts Rdm::Support::Colorize.brown(msg)
+      end
+
       def ensure_file(path_array, content = '')
         filename = File.join(*path_array)
         FileUtils.mkdir_p(File.dirname(filename))
+        return warning("File #{filename} already exists, skipping...") if File.exist?(filename)
         File.write(filename, content)
       end
 
@@ -96,6 +100,7 @@ module Rdm
         from          = filepath
         target_name ||= filepath
         to            = File.join(current_dir, package_relative_path, target_name)
+        return warning("File #{to} already exists, skipping...") if File.exist?(to)
         FileUtils.mkdir_p(File.dirname(to))
         # copy_entry(src, dest, preserve = false, dereference_root = false, remove_destination = false)
         FileUtils.copy_entry(from, to, true, false, true)
